some re-factoring, component organization and lots of vue styling to match figma specs
