+++ src/dapps/Salad/Salad.scss
@@ -1,0 +1,43 @@
+@import '~@/scss/GlobalVariables';
+%shared-text {
+    font-family: "Avenir Next";
+    font-style: normal;
+    font-weight: normal;
+}
+
+%shared-links {
+    a {
+        color: #7073b5 !important;
+    }
+}
+
+#salad-dapp-container {
+    border: 2px solid #7073b5;
+    background-color: #d0ece9;
+    box-sizing: border-box;
+    width: 715px;
+    height: auto;
+    left: 0px;
+    color: #28464F;
+    @extend %shared-text;
+    @extend %shared-links;
+    #salad-dapp-header {
+        width: 680px;
+        height: 70px;
+        margin-left: 34px;
+        .salad-dapp-title {
+            width: 188px;
+            height: 50px;
+            margin-top: 30px;
+            font-size: 42px;
+            line-height: 57px;
+            font-weight: bold !important;
+        }
+        .salad-dapp-description,
+        a {
+            margin-top: 15px;
+            font-size: 25px;
+            line-height: 34px;
+        }
+    }
+}
