+++ src/dapps/Salad/Salad.vue
@@ -1,27 +1,30 @@
 <template>
-  <div class="salad-main-container">
+  <div id="salad-dapp-container">
+    <back-button :title="$t('salad.exit-dapp')" />
+    <div id="salad-dapp-header">
+      <div class="salad-dapp-title">
+        {{ $t('salad.title') }}
+      </div>
+      <div class="salad-dapp-description">
+        {{ $t('salad.description') }} <a :href="hello">{{ $t('salad.whatisit') }}</a>
+      </div>
+    </div>
     <salad-mix></salad-mix>
   </div>
 </template>
 
 <script>
 import SaladMix from './containers/SaladMix';
+import BackButton from '@/layouts/InterfaceLayout/components/BackButton';
 
 export default {
   components: {
+    'back-button': BackButton,
     'salad-mix': SaladMix
   }
 };
 </script>
 
-<!-- Add "scoped" attribute to limit CSS to this component only -->
-<style scoped>
-.salad-main-container {
-  border: 2px solid #7070e3;
-  background-color: white;
-  border-radius: 4px;
-  grid-area: main;
-  overflow: hidden;
-  margin-bottom: 10px;
-}
+<style lang="scss" scoped>
+@import 'Salad';
 </style>
