--- src/dapps/Salad/components/SaladFooter/SaladFooter.scss
@@ -1,9 +1,0 @@
-@import '~@/scss/GlobalVariables';
-
-.salad-footer {
-  padding: 20px;
-
-  .mixAmount-label {
-    font-weight: bold;
-  }
-}
