--- src/dapps/Salad/components/SaladFooter/index.js
@@ -1,1 +1,0 @@
-export { default } from './SaladFooter';
