--- src/dapps/Salad/components/SaladHeader/SaladHeader.scss
@@ -1,6 +1,0 @@
-@import '~@/scss/GlobalVariables';
-
-.salad-header {
-  padding: 20px;
-
-}
