--- src/dapps/Salad/components/SaladHeader/index.js
@@ -1,1 +1,0 @@
-export { default } from './SaladHeader';
