+++ src/dapps/Salad/components/SaladMixFooter/SaladMixFooter.scss
@@ -1,0 +1,36 @@
+@import '~@/scss/GlobalVariables';
+#salad-mix-footer {
+    left: 237px;
+    width: 206px;
+    height: 182px;
+    margin-top: 30px;
+    margin-left: auto;
+    margin-right: auto;
+    font-family: 'Avenir Next';
+    font-size: 14px;
+    line-height: 19px;
+    color: #28464f;
+    vertical-align: middle;
+    .mix-amount-label {
+        padding-top: 20px;
+        font-weight: bold;
+        font-family: 'Avenir Next';
+        font-size: 14px;
+        line-height: 19px;
+    }
+    .mix-time-label {
+        font-family: 'Avenir Next';
+        font-size: 14px;
+        line-height: 19px;
+    }
+    a,
+    .mix-learn-more-label {
+        margin-top: 30px;
+        height: 24px;
+        color: #7073b5;
+        font-family: 'Avenir Next';
+        font-size: 18px;
+        line-height: 25px;
+        text-align: center;
+    }
+}
