+++ src/dapps/Salad/components/SaladMixFooter/SaladMixFooter.vue
@@ -1,0 +1,11 @@
+<template>
+  <div id="salad-mix-footer">
+    <div class="mix-amount-label">{{ $t('salad.mixAmount-label', ['1']) }}</div>
+    <div class="mix-time-label">{{ $t('salad.mixTime-label', ['2']) }}</div>
+    <div class="mix-learn-more-label"><a href="#">{{ $t('salad.learn-more') }}</a></div>
+  </div>
+</template>
+
+<style lang="scss" scoped>
+@import 'SaladMixFooter';
+</style>
