+++ src/dapps/Salad/components/SaladMixFooter/index.js
@@ -1,0 +1,2 @@
+export { default }
+from './SaladMixFooter';
