+++ src/dapps/Salad/components/SaladMixHeader/SaladMixHeader.scss
@@ -1,0 +1,14 @@
+@import '~@/scss/GlobalVariables';
+#salad-mix-header {
+    margin: 20px;
+    display: flex;
+    justify-content: space-between;
+    div {
+        font-family: "Avenir Next";
+        font-style: normal;
+        font-weight: normal;
+        font-size: 25px;
+        line-height: 34px;
+        color: #28464f;
+    }
+}
