+++ src/dapps/Salad/components/SaladMixHeader/SaladMixHeader.vue
@@ -1,0 +1,10 @@
+<template>
+  <div id="salad-mix-header">
+    <div>{{ $t('salad.saladHeader') }}</div>
+    <div>{{ $t('salad.nextMix-label', ['00:33:21']) }}</div>
+  </div>
+</template>
+
+<style lang="scss" scoped>
+@import 'SaladMixHeader';
+</style>
