+++ src/dapps/Salad/components/SaladMixHeader/index.js
@@ -1,0 +1,2 @@
+export { default }
+from './SaladMixHeader';
