+++ src/dapps/Salad/containers/ConfirmationForm/ConfirmationForm.vue
@@ -1,28 +1,14 @@
 <template>
   <div class="salad-confirmation-form">
-    <div class="form">
-      <b-container>
-        <salad-header></salad-header>
-        <transaction-details></transaction-details>
-
-        <b-row>
-          <div class="buttons-container">
-            <b-col>
-              <b-button class="cancel-btn" @click="cancel()">
-                {{ $t('salad.cancel-button') }}
-              </b-button>
-            </b-col>
-            <b-col>
-              <b-button :disabled="isSubmitting" class="submit-btn" @click="confirmDeposit()">
-                {{ isSubmitting ? $t('salad.pending-button') : $t('salad.confirm-button') }}
-              </b-button>
-            </b-col>
-          </div>
-        </b-row>
-        
-        <salad-footer></salad-footer>
-      </b-container>
-    </div>
+    <transaction-details></transaction-details>
+    <b-container class="buttons-container">
+      <b-button class="button cancel-btn" @click="cancel()">
+        {{ $t('salad.cancel-button') }}
+      </b-button>
+      <b-button :disabled="isSubmitting" class="button submit-btn" @click="confirmDeposit()">
+        {{ isSubmitting ? $t('salad.pending-button') : $t('salad.confirm-button') }}
+      </b-button>
+    </b-container>
   </div>
 </template>
 
@@ -33,8 +19,6 @@ import TransactionDetails from '../../components/TransactionDetails';
 
 export default {
   components: {
-    'salad-footer': SaladFooter,
-    'salad-header': SaladHeader,
     'transaction-details': TransactionDetails,
   },
   props: {
@@ -52,5 +36,5 @@ export default {
 </script>
 
 <style lang="scss" scoped>
-@import 'Confirmationform.scss';
+@import 'Confirmationform';
 </style>
