--- src/dapps/Salad/containers/SaladMixContent/SaladMixContent.scss
@@ -1,21 +1,0 @@
-@import '~@/scss/GlobalVariables';
-
-h3 {
-  margin: 40px 0 0;
-}
-ul {
-  list-style-type: none;
-  padding: 0;
-}
-li {
-  display: inline-block;
-  margin: 0 10px;
-}
-a {
-  color: black;
-}
-.salad-mix-content {
-  background-color: white !important;
-  border-radius: 25px;
-  border: 2px solid #7070e3;
-}
