--- src/dapps/Salad/containers/SaladMixContent/index.js
@@ -1,1 +1,0 @@
-export { default } from './SaladMixContent';
