+++ src/translations/dapps-salad/en_US.json
@@ -1,36 +1,37 @@
 {
-  "title": "Salad",
-  "desc": "Privacy is healthy. Make some Salad and keep your transactions to yourself. What is Salad?",
-  "saladHeader": "Send ETH Privately",
-  "txDetailsHeader": "Transaction Details",
-  "amountSent-label": "Amount Sent: {0} ETH",
-  "amountReceived-label": "Amount Received: {0} ETH",
-  "totalFees-label": "Total Fees: {0}",
-  "feeCalculation-link": "How are fees calculated?",
-  "salad-mix-popover": "Send ETH Privately",
-  "nextMix-label": "Next Mix in: {0}",
-  "deliveryAddress-ph": "e.g. 0x0343..............4343",
-  "deliveryAddress-label": "Where do you want funds delivered?",
-  "depositAddress-label": "Funds will be sent from ",
-  "depositAddress-label2": " [user's current address]",
-  "startDeposit-button": "Start Deposit",
-  "cancel-button": "Cancel",
-  "confirm-button": "Confirm",
-  "pending-button": "Pending....",
-  "startNewMix-button": "Start a New Mix",
-  "mixAmount-label": "Salad mixes {0} ETH at a time.",
-  "mixTime-label": "Funds can take up to {0} hours to reach the destination address." , 
-  "dealId-label": "TxHash: ",
-  "exit-dapp": "Exit Dapp",
-  "learn-more": "Learn More",
-  "feeCalculation-label": "How are fees calculated?",
-  "minimumEthErrMessage": "You don't have enough ETH in your account for this transaction.",
-  "saladSuccessHeader": "Success! Your salad is being tossed!",
-  "pendingStatus": "Status: Pending Deal",
-  "pendingStatusMessage": "Your deposit has been made, and you will see funds in your recipient account in approximately {0} minutes.",
-  "submittedStatus": "Status: Deal Submitted",
-  "submittedStatusMessage": "Your deposit has been made, and you will see funds in your recipient account in just a few minutes.",
-  "completedStatus": "Status: Deal Complete",
-  "completedStatusMessage": "Your deposit has been made, and you should see funds in your recipient account now.",
-  "validRecipientErrMessage": "RecipientAddress must be a valid Ethereum address."
+    "title": "Salad",
+    "description": "Privacy is healthy. Make some Salad and keep your transactions to yourself.",
+    "whatisit": "What is Salad?",
+    "saladHeader": "Send ETH Privately",
+    "txDetailsHeader": "Transaction Details",
+    "amountSent-label": "Amount Sent: {0} ETH",
+    "amountReceived-label": "Amount Received: {0} ETH",
+    "totalFees-label": "Total Fees: {0}",
+    "feeCalculation-link": "How are fees calculated?",
+    "salad-mix-popover": "Send ETH Privately",
+    "nextMix-label": "Next Mix in: {0}",
+    "deliveryAddress-ph": "e.g. 0x0343..............4343",
+    "deliveryAddress-label": "Where do you want funds delivered?",
+    "depositAddress-label": "Funds will be sent from ",
+    "depositAddress-label2": " [user's current address]",
+    "startDeposit-button": "Start Deposit",
+    "cancel-button": "Cancel",
+    "confirm-button": "Confirm",
+    "pending-button": "Pending....",
+    "startNewMix-button": "Start a New Mix",
+    "mixAmount-label": "Salad mixes {0} ETH at a time.",
+    "mixTime-label": "Funds can take up to {0} hours to reach the destination address.",
+    "dealId-label": "TxHash: ",
+    "exit-dapp": "Exit Dapp",
+    "learn-more": "Learn More",
+    "feeCalculation-label": "How are fees calculated?",
+    "minimumEthErrMessage": "You don't have enough ETH in your account for this transaction.",
+    "saladSuccessHeader": "Success! Your salad is being tossed!",
+    "pendingStatus": "Status: Pending Deal",
+    "pendingStatusMessage": "Your deposit has been made, and you will see funds in your recipient account in approximately {0} minutes.",
+    "submittedStatus": "Status: Deal Submitted",
+    "submittedStatusMessage": "Your deposit has been made, and you will see funds in your recipient account in just a few minutes.",
+    "completedStatus": "Status: Deal Complete",
+    "completedStatusMessage": "Your deposit has been made, and you should see funds in your recipient account now.",
+    "validRecipientErrMessage": "RecipientAddress must be a valid Ethereum address."
 }
